feat: Sell tickets from travel search for train in zone a
