+++ src/travel-details-screens/components/use-can-sell-collab-ticket.ts
@@ -1,0 +1,24 @@
+import {TripPattern} from '@atb/api/types/trips';
+import {APP_ORG} from '@env';
+import {
+  Mode,
+  TariffZone,
+} from '@atb/api/types/generated/journey_planner_v3_types';
+
+export function useCanSellCollabTicket(tripPattern: TripPattern) {
+  const someLegsByTrain = someLegsAreByTrain(tripPattern);
+  const tariffZonesHaveZoneA = (tariffZones?: TariffZone[]) =>
+    tariffZones?.some((a) => a.id === 'ATB:TariffZone:1');
+  const allLegsInZoneA = tripPattern.legs
+    .filter((a) => a.mode !== Mode.Foot)
+    .every(
+      (a) =>
+        tariffZonesHaveZoneA(a.fromPlace.quay?.tariffZones) &&
+        tariffZonesHaveZoneA(a.toPlace.quay?.tariffZones),
+    );
+  return someLegsByTrain && allLegsInZoneA && APP_ORG === 'atb';
+}
+
+function someLegsAreByTrain(tripPattern: TripPattern): boolean {
+  return tripPattern.legs.some((leg) => leg.mode === Mode.Rail);
+}
